+++ src/app/oqapi/query-form/oqt-api-query-form/oqt-api-query-form.component.ts
@@ -9,7 +9,7 @@ import { SuiMultiSelectSearchDropdownComponent } from '../../../shared/component
 import { PrismEditorComponent } from '../../../shared/components/prism-editor/prism-editor.component';
 import { SimpleIndicatorComponent } from './simple-indicator/simple-indicator.component';
 import { AttributeCompletenessAttributesComponent } from './attribute-completeness-attributes/attribute-completeness-attributes.component';
-import { ThematicalAccuracyIndicatorComponent } from './thematical-accuracy-indicator/thematical-accuracy-indicator.component';
+import { ThematicAccuracyIndicatorComponent } from './thematic-accuracy-indicator/thematic-accuracy-indicator.component';
 import { KeyValuePipe } from '@angular/common';
 
 @Component({
@@ -17,7 +17,7 @@ import { KeyValuePipe } from '@angular/common';
     templateUrl: './oqt-api-query-form.component.html',
     styleUrls: ['./oqt-api-query-form.component.css'],
     viewProviders: [{ provide: ControlContainer, useExisting: NgForm }],
-    imports: [FormsModule, SuiMultiSelectSearchDropdownComponent, PrismEditorComponent, SimpleIndicatorComponent, AttributeCompletenessAttributesComponent, ThematicalAccuracyIndicatorComponent, KeyValuePipe]
+  imports: [FormsModule, SuiMultiSelectSearchDropdownComponent, PrismEditorComponent, SimpleIndicatorComponent, AttributeCompletenessAttributesComponent, KeyValuePipe, ThematicAccuracyIndicatorComponent]
 })
 export class OqtApiQueryFormComponent implements OnInit, OnDestroy {